fix(schema) do not produce incomplete types when auto-converting legacy "table"
When auto-converting a legacy schema from a plugin, we need to determine
if something of type "table" is an array, map, set or record. When
we failed to pin this down, we were producing an incomplete schema,
which led to runtime crashes.

if we have a table type that wasn't further specified in any way,
we now make it a map from strings to strings, which is the most generic
type (apart from abusing "any", since we don't want dynamically-typed
plugin schemas).