fix(schema) verify the validity of auto-converted schemas
This is a defensive measure. The best-effort legacy schema converter
should either produce a valid schema or report failure. In case it
misbehaves, the MetaSchema validator should be able to catch that,
so that an invalid schema does not slip through.

Also fixes the formatting of an error message when reporting on an invalid
schema.

See #4392.